Change X and Y coordinates to double

+++ src/main/java/group22/space_invaders/unit/Alien.java
@@ -8,7 +8,7 @@ package group22.space_invaders.unit;
 
 public class Alien extends Unit {
 	
-	public Alien(float X, float Y){
+	public Alien(double X, double Y){
 		super(X,Y);
 	}
