+++ src/test/java/group22/space_invaders/unit/AlienBulletTest.java
@@ -8,7 +8,7 @@ package group22.space_invaders.unit;
 public class AlienBulletTest extends BulletTest{
 
 	@Override
-	public Unit createInstance(float X, float Y) {
+	public Unit createInstance(double X, double Y) {
 		return new AlienBullet(X,Y);
 	}
