+++ src/test/java/group22/space_invaders/unit/AlienTest.java
@@ -11,7 +11,7 @@ import org.junit.Test;
 public class AlienTest extends UnitTest{
 
 	@Override
-	public Unit createInstance(float X, float Y) {
+	public Unit createInstance(double X, double Y) {
 		return new Alien(X,Y);
 	}
